Access user accounts using /etc/passwd,/etc/shadow
The server will now be able to access user accounts using
passwd file. This will be handled in login.c file. If the
user name doesn't exist then the client will be disconnected.
The same will happend if the password is incorrect. New library
was added in the make file 'crypt'. This is need to verify that
the password is correct.