Added login functionality. Updated testcases
The server are now able to accept logins. The username
and passwd will be matched from /etc/passwd and /etc/shadow
files.
Changed the test-cases so it uses the right account otherwise
they will not be able to login. The testcases need an account
with username: 'tester' and passwd: 'celeron'.
The server must now be started with the sudo command otherwise
it will not be able to read from /etc/passwd file.